+++ changelog.md
@@ -7,6 +7,7 @@
 
 ### Improvements
 - Player count and roster errors from TTS boxscore upload are passed to Discord instead of blocking the upload
+- Improved the #results message to link to thre game thread when possible
 
 ### Bug Fixes
 - Subscribing to a scheduling poll now notifies on every new "Yes" answer instead of just poll close.